test: split PDF visual overlay retention coverage
